Externalize onStorage and deleteOldChages
